don't send new watch command before poling
* don't rescue exception: http://stackoverflow.com/questions/10048173/why-is-it-bad-style-to-rescue-exception-e-in-ruby

* improve socket reading:
Since we already started the watch with gpsd_client.start there is no need to issue a new ?WATCH command to create a new listener when .get_position was called.
The issue that occurs here is that each watch creates a couple of lines followed by poll which again creates a line.
gets only reads one line at a time, leaving data on the socket which was no read.
If sleeping for let's say 10 sec, a poll will deliver data from 10 sec ago.
After some time the data on the socket adds up and the client crashes.